+++ electron.vite.config.ts
@@ -1,0 +1,29 @@
+import { resolve } from 'node:path'
+import { defineConfig } from 'electron-vite'
+import react from '@vitejs/plugin-react'
+
+export default defineConfig({
+  main: {
+    build: {
+      rollupOptions: {
+        input: resolve(__dirname, 'src/electron/main/index.ts')
+      }
+    }
+  },
+  preload: {
+    build: {
+      rollupOptions: {
+        input: resolve(__dirname, 'src/electron/preload/index.ts')
+      }
+    }
+  },
+  renderer: {
+    root: resolve(__dirname, 'src/electron/renderer'),
+    plugins: [react()],
+    build: {
+      rollupOptions: {
+        input: resolve(__dirname, 'src/electron/renderer/index.html')
+      }
+    }
+  }
+})